Refactor: Split exposure export utils into modules
Restructure the exposure export utility functions into smaller, more focused modules for improved code organization and maintainability.
